% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auto}
\alias{auto}
\title{Auto.}
\format{
A data frame with 3 variables:
\describe{
\item{\code{year}}{ household number }
\item{\code{invest}}{ investment}
\item{\code{vsh}}{ value of outstanding shares } }
}
\usage{
auto
}
\description{
Data on investment and value of outstanding shares for the years 1935-1953
 in Econometrics_Stat3061 module.
}
\keyword{datasets}
