% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consmption}
\alias{consmption}
\title{Consmption.}
\format{
A data frame with 3 variables:
\describe{
\item{\code{household}}{ household number }
\item{\code{income}}{ household income }
\item{\code{expen}}{ household consumption expenditure } }
}
\usage{
consmption
}
\description{
Consmption dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
