% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cps2}
\alias{cps2}
\title{cps2.}
\format{
A data frame with ten variables:
\describe{
\item{\code{wage}}{  earnings per hour }
\item{\code{educ}}{  years of education}
\item{\code{exper}}{ experience }
\item{\code{female}}{  =1 if female }
\item{\code{black}}{  =1 if black }
\item{\code{married}}{  =1 if married }
\item{\code{union}}{  =1 if union member }
\item{\code{south}}{  =1 if south }
\item{\code{fulltime}}{  =1 if full time worker }
\item{\code{metro}}{  =1 if lives in metropolitan area}}
}
\usage{
cps2
}
\description{
wage dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
