% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo2}
\alias{demo2}
\title{Demo2.}
\format{
A data frame with 2 variables:
\describe{
\item{\code{x}}{ weekly income }
\item{\code{y}}{ weekly consumption expenditure } }
}
\usage{
demo2
}
\description{
hypothetical  data,  a dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
