% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ_inc}
\alias{educ_inc}
\title{educ_inc.}
\format{
A data frame with six variables:
\describe{
\item{\code{faminc}}{ Family income in 2006 dollars}
\item{\code{hedu}}{ Husband's educational attainment, in years}
\item{\code{wedu}}{ Wife's educational attainment, in years}
\item{\code{k16}}{  Number of children less than 6 years old in household}
\item{\code{x5}}{x5 is an artificially generated variable}
\item{\code{x6}}{ x6 is an artificially generated variable} }
}
\usage{
educ_inc
}
\description{
Wage  data,  a dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
