% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ele_cons}
\alias{ele_cons}
\title{ele_cons.}
\format{
A data frame with three variables: \code{year}, \code{X}, \code{Y}
}
\usage{
ele_cons
}
\description{
data on the percentage rate of change in electricity consumption (millions KWH)
(Y) and the rate of change in the price of electricity (Birr/KWH) (X) for the
years 1979 – 1994,  a dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
