% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food}
\alias{food}
\title{Food}
\format{
A data frame with two variables: \code{food_exp}, \code{income}
}
\usage{
food
}
\description{
Dataset in  Econometrics_Stat3061 module.
}
\keyword{datasets}
