% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_cons}
\alias{food_cons}
\title{Food_cons}
\format{
A data frame with four variables: \code{Year}, \code{Y},  \code{X2} and \code{X3}
}
\usage{
food_cons
}
\description{
data on per capita food consumption (Y), price of food (X2 ) and
per capita income (X3) where retail price of food and per capita disposable income are deflated by
the Consumer Price Index.
}
\keyword{datasets}
