% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_price}
\alias{house_price}
\title{House_price.}
\format{
A data frame with seven variables:
\describe{
\item{\code{price }}{  house price, in \$1000}
\item{\code{sqft}}{ square feet of living area, in 100s}
\item{\code{age}}{ house age, in years}
\item{\code{utown}}{  =1 if close to university}
\item{\code{pool}}{ =1 if house has pool}
\item{\code{fplace}}{=1 if house has fireplace}}
}
\usage{
house_price
}
\description{
Data on some characteristics of the wine industry in Australia, a dataset
for simultaneous equation Models(SEM) in Econometrics_Stat3061 module.
}
\keyword{datasets}
