% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{import}
\alias{import}
\title{Import.}
\format{
An object of class \code{data.frame} with 19 rows and 5 columns.
}
\usage{
import
}
\description{
Data set of imports in Econometrics_Stat3061 module.
}
\details{
@format A data frame with five variables: \code{year}, \code{imports},
  \code{GDP}, \code{stock.formation} and \code{consumption}
}
\keyword{datasets}
