% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{savingdata}
\alias{savingdata}
\title{Saving data}
\format{
A data frame with three variables: \code{year}, \code{saving},and \code{GDP}
}
\usage{
savingdata
}
\description{
Data is on domestic household savings and gross domestic product  for India
for the period 1980 to 2002, a dataset in Econometrics_Stat3061 module.
}
\keyword{datasets}
