% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine.}
\format{
A data frame with seven variables:
\describe{
\item{\code{year}}{age in years}
\item{\code{consumption}}{real per capita consumption of wine}
\item{\code{storage_costs}}{index of storage costs}
\item{\code{price_wine}}{ price of wine relative to CPI}
\item{\code{price_beer}}{price of beer relative to CPI}
\item{\code{advertising}}{real per capita advertising expenditure}
\item{\code{income}}{real per capita disposable income}}
}
\usage{
wine
}
\description{
Data on some characteristics of the wine industry in Australia, a dataset
for simultaneous equation Models(SEM) in Econometrics_Stat3061 module.
}
\keyword{datasets}
